Solve token verification and chat problems
